refactor(e2e): extract shared mocks and helpers to mock-helpers.ts
Move SUPABASE_STORAGE_KEY, createFakeSession, setupEditorMocks,
and gotoEditor into e2e/utils/mock-helpers.ts. Both anonymous-flow
and accessibility specs now import from the shared utility instead
of duplicating 130+ lines of mock setup.
